core/requirements.txt
# Core
flask>=3.0.0
flask-cors>=4.0.0
flask-restx>=1.3.0
python-dotenv>=1.0.0
PyJWT>=2.8.0

# YouTube Music
ytmusicapi>=1.4.0
yt-dlp>=2024.1.0

# Database
flask-sqlalchemy>=3.1.0

# Caching
redis>=5.0.0

# HTTP
requests>=2.31.0
certifi>=2024.0.0

# Audio Processing
faster-whisper>=0.10.0

# Rate Limiting
flask-limiter>=3.5.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0

# Production
gunicorn>=21.0.0

# Documentation
mkdocs-material>=9.5.0
